Assignment 6 Completed (#60)
* Assignment 2 Completed

* Update README.md

README re-written to explain application more succinctly.

* temp save for work

* Completed Assignment 3

* Fixed incorrect answer for Assignment 3

* Update README.md

Final touches for ReadMe

* Added increment to sameSite variable

* Partially completed assignment 4

* Update README.md

README.md updated for better formatting

* Completed Assignment 5

* Clean ups following videos for Assignment 5

* Added README.md and screenshots

* completed assignment 6

* Add README for Assignment 6 on browser fingerprinting

Detailed the implementation and testing of an Express server for browser fingerprinting.

* Update README for Assignment 6

Added a YouTube link and removed duplicate content.

* Assignment 7 completion

* Add README for Assignment 7 

This README outlines the details of Assignment 7, including the task of simulating a phishing login page and the methods used for user data exfiltration.